fix: remove timeouts from pubsub tests
Instead of waiting for 800ms for pubsub subscriptions to propagate,
wait for the publisher to see the reviever in it's list of subscribers
for a given topic.

This is because sometimes in CI it can taker longer than 800ms for
the subs list to update which causes the tests to be flaky.