fix: start daemons in series
We store references to the daemon client in the return value of
`Promise.all` and use them to stop the daemons after a test run -
if one or more daemons fail to start, we lose all references to
daemons that started successfully so cannot stop them and the test
run hangs.

Instead start them individually so we can tear down any that started
successfully.